function extracted, looped through, and dom renders
