+++ index.html
@@ -13,21 +13,7 @@
         <h1>Photo Gallery</h1>
     </header>
     <main>
-        <ul id="image-list">
-            <li>
-                <h2>Baby Rhino</h2>
-                <img src="https://secure.img1-ag.wfcdn.com/im/17007094/resize-h800%5Ecompr-r85/3589/35892451/Baby+Rhino+Figurine.jpg">
-            </li>
-            <li>
-                <h2>Cera</h2>
-                <img src="https://vignette.wikia.nocookie.net/landbeforetime/images/c/c3/Cera_infobox.png/revision/latest?cb=20180422005003">
-            </li>
-            <li>
-                <h2>markhor</h2>
-                <img src="https://upload.wikimedia.org/wikipedia/commons/thumb/3/38/Markhor_Schraubenziege_Capra_falconeri_Zoo_Augsburg-02.jpg/220px-Markhor_Schraubenziege_Capra_falconeri_Zoo_Augsburg-02.jpg">
-            </li>
-
-        </ul>
+        <ul id="image-list"></ul>
     </main>
     <footer></footer>
