+++ make-image-template.js
@@ -1,0 +1,14 @@
+function makeImageTemplate(image) {
+    const html = /*html*/`
+        <li>
+            <h2>${image.title}</h2>
+            <img src="${image.url}">
+        </li>
+    `;
+
+    const template = document.createElement('template');
+    template.innerHTML = html;
+    return template.content;
+}
+
+export default makeImageTemplate;
