+++ style.css
@@ -2,30 +2,26 @@
     background: silver;
 }
 
-#image-list ul {
+#image-list {
     display: flex;
-    width: 100vh;
+    flex-wrap: wrap;
+    padding: 0;
 }
 
 #image-list li {
-    display: flex;
-    justify-content: center;
-    width: 80vw;
-    height: 75%;
-    flex-direction: row;
+    list-style: none;
+    position: relative;
+    width: 400px;
     border: 1px solid black;
     margin: 10px;
 }
 
 #image-list h2 {
-    display: flex;
-    width: 10vw;
-    justify-content: flex-start;
+
 }
 
 #image-list img {
-    display: flex;
-    width: auto;
-    padding: 10px;
+    width: 90%;
+    /* padding: 10px; */
     border-radius: 45px;
 }
